rename: 폴더 구조 변경 및 viewController 이름 변경
